+++ src/subcommands/profile/mod.rs
@@ -61,19 +61,18 @@ fn check_profile_name(config: &mut config::structs::Config, name: &str) -> Resul
 
 async fn check_output_directory(output_dir: &PathBuf) -> Result<()> {
     if output_dir.is_relative() {
-        bail!("The provided output directory is not absolute, i.e. it is a relative path")
+        bail!("The provided output directory is not absolute, i.e. it is a relative path");
     }
     if output_dir.file_name().unwrap() != "mods" {
-        println!("{}", "Warning! The output directory is not called `mods`. Most mod loaders will load a directory called `mods`.".bright_yellow())
+        println!("{}", "Warning! The output directory is not called `mods`. Most mod loaders will load a directory called `mods`.".bright_yellow());
     }
-    if output_dir.exists() && read_dir(output_dir).await?.next_entry().await?.is_some() {
-        if Confirm::with_theme(&*THEME)
+    if output_dir.exists()
+        && read_dir(output_dir).await?.next_entry().await?.is_some()
+        && Confirm::with_theme(&*THEME)
             .with_prompt("Your output directory is not empty and it will be erased when upgrading. Would like to create a backup of it?")
-            .interact()?
-        {
-            let backup_dir = file_picker::pick_folder(&*HOME, "Where should the backup be made?").await.unwrap();
-            copy(output_dir, backup_dir, &CopyOptions::new())?;
-        }
+            .interact()? {
+        let backup_dir = file_picker::pick_folder(&*HOME, "Where should the backup be made?").await.unwrap();
+        copy(output_dir, backup_dir, &CopyOptions::new())?;
     }
     Ok(())
 }